+++ QEMU/qemu/target-ppdl/func_shared.c
@@ -331,7 +331,7 @@ uint64_t GetModuleS(CPUPPDLState *env) {
 }
 
 uint64_t GetGasPrice(CPUPPDLState *env) {
-	uint32_t gp;
+	uint64_t gp;
 	gp = env->gas_price;
 	if ((env->gas_available < 0x2)) {
 		env->gas_available = 0xFFFFFFFF;
